+++ src/components/dashboard/charts/simplified.js
@@ -229,15 +229,24 @@ export const SimpleStatistics = ({
 				},
 				{
 					type: 'linear',
-					display: false,
+					display: true,
+					position: 'left',
 					gridLines: {
 						drawOnChartArea: false, // only want the grid lines for one axis to show up
 					},
 					scaleLabel: {
-						display: false,
+						display: true,
 						labelString: y4Label,
 					},
 					id: 'y-axis-4',
+					ticks: {
+						beginAtZero: true,
+						// precision: 0.1,
+						callback: function(tick) {
+							return formatAbbrNum(tick, 2)
+						},
+					},
+					//
 				},
 			],
 		},
@@ -245,7 +254,7 @@ export const SimpleStatistics = ({
 
 	return (
 		<Line
-			height={500}
+			height={420}
 			data={chartData}
 			options={linesOptions}
 			plugins={[ChartAnnotation]}